Add Thermal Conductivity Calculating!!!

+++ Core/Inc/Menus.h
@@ -15,9 +15,9 @@ extern int screen;           // Current screen
 extern int button;           // Last button pressed
 extern int number;           // Number input from numberSelect
 extern double deltaTemp;     // Current temperature for test
-extern double deltaTime;     // Current time for test
+extern int deltaTime;     // Current time for test
 extern double deltaTempDefault;  // Default temperature
-extern double deltaTimeDefault;  // Default duration
+extern int deltaTimeDefault;  // Default duration
 extern int brightness;       // Display brightness
 extern int both;             // Flag for "both" option in test
 extern int heater;
